Added logic for unauthorized users

+++ ecommerce/store/models.py
@@ -14,7 +14,7 @@ class Customer(models.Model):
 
 class Product(models.Model):
     name = models.CharField(max_length=200, null=True)
-    price = models.FloatField()
+    price = models.DecimalField(max_digits=7, decimal_places=2)
     digital = models.BooleanField(default=False, null=True, blank=False)
     image = models.ImageField(null=True, blank=True)
 
@@ -39,7 +39,7 @@ class Order(models.Model):
 
     def __str__(self):
         return str(self.id)
-    
+
     @property
     def shipping(self):
         shipping = False
